+++ src/mindlessgen/molecules/postprocess.py
@@ -22,8 +22,20 @@ def postprocess_mol(
     """
 
     # Run a singlepoint calculation with the postprocess engine to see if it is stable
-    try:
-        engine.singlepoint(mol, verbosity=verbosity)
-    except RuntimeError as e:
-        raise RuntimeError("Single point calculation in postprocessing failed.") from e
-    return mol
+    if verbosity > 2:
+        print("Postprocessing molecule...")
+    if config.optimize:
+        try:
+            postprocmol = engine.optimize(
+                mol, max_cycles=config.opt_cycles, verbosity=verbosity
+            )
+        except RuntimeError as e:
+            raise RuntimeError("Optimization in postprocessing failed.") from e
+    else:
+        try:
+            engine.singlepoint(mol, verbosity=verbosity)
+        except RuntimeError as e:
+            raise RuntimeError(
+                "Single point calculation in postprocessing failed."
+            ) from e
+    return postprocmol