+++ test/test_qm/test_xtb.py
@@ -8,6 +8,7 @@ import pytest
 import numpy as np
 from mindlessgen.qm import XTB, get_xtb_path  # type: ignore
 from mindlessgen.molecules import Molecule  # type: ignore
+from mindlessgen.prog import XTBConfig  # type: ignore
 
 
 # mark all tests as optional as they depend on the availability of xtb
@@ -17,9 +18,10 @@ def test_xtb_optimize_xtb(coordinates_ethanol: np.ndarray) -> None:
     """
     Test the optimization of ethanol with xtb.
     """
+    cfg = XTBConfig()
     xtb_path = get_xtb_path()
     if xtb_path:
-        xtb = XTB(xtb_path)
+        xtb = XTB(xtb_path, cfg)
     else:
         raise RuntimeError("xtb not found.")
     mol = Molecule()
@@ -114,13 +116,14 @@ def test_check_gap_low_gap(mol_C2H4N1O1Au1: Molecule, mol_H3B4Pd1Rn1: Molecule):
     Test check_gap with a molecule that has a low HOMO-LUMO gap and one with a sufficient gap.
     """
 
+    cfg = XTBConfig()
     try:
         xtb_path = get_xtb_path()
         if not xtb_path:
             raise ImportError("xtb not found.")
     except ImportError as e:
         raise ImportError("xtb not found.") from e
-    engine = XTB(xtb_path)
+    engine = XTB(xtb_path, cfg)
     # Test for molecule with low gap
     result_low_gap = engine.check_gap(mol_H3B4Pd1Rn1, threshold=0.5)
     assert result_low_gap is False